Fixing The Sonar Cloud

+++ source/core/src/main/com/deco2800/game/components/mainmenu/MainMenuDisplay.java
@@ -157,7 +157,7 @@ public class MainMenuDisplay extends UIComponent {
         /**
          * Sets the size and position of the button after texture applied.
          */
-        settingsBtn.setBounds(centreWidth-width35Percent,centreHeight-height83Percent,
+        settingsBtn.setBounds((float)(centreWidth-width35Percent),(float)(centreHeight-height83Percent),
                 buttonDimensionsWidth, buttonDimensionsHeight);
 
 
@@ -172,7 +172,7 @@ public class MainMenuDisplay extends UIComponent {
         /**
          * Sets the size and position of the button after texture applied.
          */
-        exitBtn.setBounds(centreWidth,centreHeight-height98Percent,
+        exitBtn.setBounds(centreWidth,(float)(centreHeight-height98Percent),
                 buttonDimensionsWidth, buttonDimensionsHeight);
 
         /**
@@ -200,7 +200,7 @@ public class MainMenuDisplay extends UIComponent {
         /**
          * Sets the size and position of the button after texture applied, for Mute and Currently Muted both.
          */
-        muteBtn.setBounds(centreWidth+width35Percent,centreHeight-height83Percent,
+        muteBtn.setBounds((float)(centreWidth+width35Percent),(float)(centreHeight-height83Percent),
                 buttonDimensionsWidth, buttonDimensionsHeight);
 
         /**
@@ -214,7 +214,7 @@ public class MainMenuDisplay extends UIComponent {
         /**
          * Sets the size and position of the button after texture applied.
          */
-        leaderBoardBtn.setBounds(centreWidth+width35Percent,centreHeight-height53Percent,
+        leaderBoardBtn.setBounds((float)(centreWidth+width35Percent),(float)(centreHeight-height53Percent),
                 buttonDimensionsWidth, buttonDimensionsHeight);
 
         TextButton levelEditorBtn = new TextButton("LevelEditor", skin);
